requirements.txt
Django==2.2.19
django-filter==2.1.0
pytz==2021.3
sqlparse==0.4.2
